backend/requirements.txt
flask
flask-cors
groq
python-dotenv
google-generativeai
Pillow